Updated Tiles - Created Units Package
Created new package for units. This will be similar to Tiles, where a
single parent class defines what a unit is at a basic level, and each
individual unit will specify it's own behavior. This is needed early on
as I intend to use the tiles to hold a representation of the overall
board, not the players.